requirements.txt
streamlit
time
streamlit_chat
PyPDF2
langchain_community.embeddings
langchain.text_splitter
langchain_community.vectorstores
langchain.chains.question_answering
langchain_openai
langchain.prompts
langchain.chains.retrieval_qa.base
langchain_core.prompts
langchain_openai
langchain_core.runnables
langchain_core.output_parsers
openai
dotenv